Moving all non-public packages to internal (#401)
