+++ clients/ui/bff/internal/api/healthcheck_handler.go
