+++ clients/ui/bff/internal/api/model_versions_handler.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"github.com/julienschmidt/httprouter"
 	"github.com/kubeflow/model-registry/pkg/openapi"
-	"github.com/kubeflow/model-registry/ui/bff/integrations"
-	"github.com/kubeflow/model-registry/ui/bff/validation"
+	"github.com/kubeflow/model-registry/ui/bff/internal/integrations"
+	"github.com/kubeflow/model-registry/ui/bff/internal/validation"
 	"net/http"
 )
