+++ clients/ui/bff/internal/api/registered_models_handler.go
@@ -4,12 +4,12 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/kubeflow/model-registry/ui/bff/internal/integrations"
+	"github.com/kubeflow/model-registry/ui/bff/internal/validation"
 	"net/http"
 
 	"github.com/julienschmidt/httprouter"
 	"github.com/kubeflow/model-registry/pkg/openapi"
-	"github.com/kubeflow/model-registry/ui/bff/integrations"
-	"github.com/kubeflow/model-registry/ui/bff/validation"
 )
 
 type RegisteredModelEnvelope Envelope[*openapi.RegisteredModel, None]
