+++ clients/ui/bff/internal/data/health_check_model.go
