+++ clients/ui/bff/internal/data/model_registry_client.go
