+++ clients/ui/bff/internal/data/model_version.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/kubeflow/model-registry/pkg/openapi"
-	"github.com/kubeflow/model-registry/ui/bff/integrations"
+	"github.com/kubeflow/model-registry/ui/bff/internal/integrations"
 	"net/url"
 )
