+++ clients/ui/bff/internal/mocks/static_data_mock.go
